fix(alacritty): set TERM variable properly
Some programs and their dependencies (peco) do not support the `$TERM`
value of `alacritty`, and as such causes major bugs. In the past, we had
been solving this by overriding the TERM env var in .zshrc if the TERM
was currently "alacritty" and we had peco installed.

Instead, now use alacritty.yml to set the TERM variable every time. This
removes clutter from the zshrc and isolates the solution to just the
terminal (alacritty) where the problem occurs